Add IT for UserRestController

+++ taskmaster/src/main/java/bg/softuni/taskmaster/web/controller/UserRestController.java
@@ -5,6 +5,7 @@ import bg.softuni.taskmaster.service.AuthorizationService;
 import bg.softuni.taskmaster.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -15,27 +16,32 @@ public class UserRestController {
     private final UserService userService;
     private final AuthorizationService authorizationService;
 
-    @GetMapping("/{id}")
-    public ResponseEntity<UserInfoDTO> userInfo(@PathVariable Long id) {
-        return ResponseEntity.ok(userService.getInfo(id));
-    }
-
     @DeleteMapping("/{id}")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
         if (!userService.exists(id)) {
             return ResponseEntity.badRequest().build();
         }
+        userService.delete(id);
         return ResponseEntity.ok().build();
     }
 
     @PatchMapping("/make-admin/{id}")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Void> makeAdmin(@PathVariable Long id) {
+        if (!userService.exists(id)) {
+            return ResponseEntity.badRequest().build();
+        }
         authorizationService.makeAdmin(id);
         return ResponseEntity.ok().build();
     }
 
     @PatchMapping("/remove-admin/{id}")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Void> removeAdmin(@PathVariable Long id) {
+        if (!userService.exists(id)) {
+            return ResponseEntity.badRequest().build();
+        }
         authorizationService.removeAdmin(id);
         return ResponseEntity.ok().build();
     }
